+++ packages/components/src/index.js
@@ -45,3 +45,4 @@ export { default as OverlayHeader } from './overlayHeader';
 export { default as Image } from './image';
 export { default as Markdown } from './markdown';
 export { default as Card } from './card';
+export { default as Summary } from './summary';